+++ changelog.md
@@ -1,2 +1,4 @@
 v1.1
-- release pertama
+
+* Updated to NVDA 2026.1, and updated addon templates.
+